+++ example/lib/main.dart
@@ -3,6 +3,8 @@ import 'dart:ui';
 import 'dart:async';
 import 'dart:io';
 
+import 'package:device_info/device_info.dart';
+import 'package:android_path_provider/android_path_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:flutter_downloader/flutter_downloader.dart';
@@ -36,11 +38,11 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget with WidgetsBindingObserver {
-  final TargetPlatform platform;
+  final TargetPlatform? platform;
 
-  MyHomePage({Key key, this.title, this.platform}) : super(key: key);
+  MyHomePage({Key? key, this.title, this.platform}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => new _MyHomePageState();
@@ -106,11 +108,11 @@ class _MyHomePageState extends State<MyHomePage> {
     }
   ];
 
-  List<_TaskInfo> _tasks;
-  List<_ItemHolder> _items;
-  bool _isLoading;
-  bool _permissionReady;
-  String _localPath;
+  List<_TaskInfo>? _tasks;
+  late List<_ItemHolder> _items;
+  late bool _isLoading;
+  late bool _permissionReady;
+  late String _localPath;
   ReceivePort _port = ReceivePort();
 
   @override
@@ -145,18 +147,16 @@ class _MyHomePageState extends State<MyHomePage> {
       if (debug) {
         print('UI Isolate Callback: $data');
       }
-      String id = data[0];
-      DownloadTaskStatus status = data[1];
-      int progress = data[2];
+      String? id = data[0];
+      DownloadTaskStatus? status = data[1];
+      int? progress = data[2];
 
-      if (_tasks != null && _tasks.isNotEmpty) {
-        final task = _tasks.firstWhere((task) => task.taskId == id);
-        if (task != null) {
-          setState(() {
-            task.status = status;
-            task.progress = progress;
-          });
-        }
+      if (_tasks != null && _tasks!.isNotEmpty) {
+        final task = _tasks!.firstWhere((task) => task.taskId == id);
+        setState(() {
+          task.status = status;
+          task.progress = progress;
+        });
       }
     });
   }
@@ -172,7 +172,7 @@ class _MyHomePageState extends State<MyHomePage> {
           'Background Isolate Callback: task ($id) is in status ($status) and process ($progress)');
     }
     final SendPort send =
-        IsolateNameServer.lookupPortByName('downloader_send_port');
+        IsolateNameServer.lookupPortByName('downloader_send_port')!;
     send.send([id, status, progress]);
   }
 
@@ -180,7 +180,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return new Scaffold(
       appBar: new AppBar(
-        title: new Text(widget.title),
+        title: new Text(widget.title!),
       ),
       body: Builder(
           builder: (context) => _isLoading
@@ -198,7 +198,7 @@ class _MyHomePageState extends State<MyHomePage> {
           padding: const EdgeInsets.symmetric(vertical: 16.0),
           children: _items
               .map((item) => item.task == null
-                  ? _buildListSection(item.name)
+                  ? _buildListSection(item.name!)
                   : DownloadItem(
                       data: item,
                       onItemClick: (task) {
@@ -209,7 +209,7 @@ class _MyHomePageState extends State<MyHomePage> {
                           }
                         });
                       },
-                      onAtionClick: (task) {
+                      onActionClick: (task) {
                         if (task.status == DownloadTaskStatus.undefined) {
                           _requestDownload(task);
                         } else if (task.status == DownloadTaskStatus.running) {
@@ -255,11 +255,7 @@ class _MyHomePageState extends State<MyHomePage> {
               ),
               FlatButton(
                   onPressed: () {
-                    _checkPermission().then((hasGranted) {
-                      setState(() {
-                        _permissionReady = hasGranted;
-                      });
-                    });
+                    _retryRequestPermission();
                   },
                   child: Text(
                     'Retry',
@@ -273,46 +269,69 @@ class _MyHomePageState extends State<MyHomePage> {
         ),
       );
 
+  Future<void> _retryRequestPermission() async {
+    final hasGranted = await _checkPermission();
+
+    if (hasGranted) {
+      await _prepareSaveDir();
+    }
+
+    setState(() {
+      _permissionReady = hasGranted;
+    });
+  }
+
   void _requestDownload(_TaskInfo task) async {
     task.taskId = await FlutterDownloader.enqueue(
-        url: task.link,
-        headers: {"auth": "test_for_sql_encoding"},
-        savedDir: _localPath,
-        showNotification: true,
-        openFileFromNotification: true);
+      url: task.link!,
+      headers: {"auth": "test_for_sql_encoding"},
+      savedDir: _localPath,
+      showNotification: true,
+      openFileFromNotification: true,
+      saveInPublicStorage: true,
+    );
   }
 
   void _cancelDownload(_TaskInfo task) async {
-    await FlutterDownloader.cancel(taskId: task.taskId);
+    await FlutterDownloader.cancel(taskId: task.taskId!);
   }
 
   void _pauseDownload(_TaskInfo task) async {
-    await FlutterDownloader.pause(taskId: task.taskId);
+    await FlutterDownloader.pause(taskId: task.taskId!);
   }
 
   void _resumeDownload(_TaskInfo task) async {
-    String newTaskId = await FlutterDownloader.resume(taskId: task.taskId);
+    String? newTaskId = await FlutterDownloader.resume(taskId: task.taskId!);
     task.taskId = newTaskId;
   }
 
   void _retryDownload(_TaskInfo task) async {
-    String newTaskId = await FlutterDownloader.retry(taskId: task.taskId);
+    String? newTaskId = await FlutterDownloader.retry(taskId: task.taskId!);
     task.taskId = newTaskId;
   }
 
-  Future<bool> _openDownloadedFile(_TaskInfo task) {
-    return FlutterDownloader.open(taskId: task.taskId);
+  Future<bool> _openDownloadedFile(_TaskInfo? task) {
+    if (task != null) {
+      return FlutterDownloader.open(taskId: task.taskId!);
+    } else {
+      return Future.value(false);
+    }
   }
 
   void _delete(_TaskInfo task) async {
     await FlutterDownloader.remove(
-        taskId: task.taskId, shouldDeleteContent: true);
+        taskId: task.taskId!, shouldDeleteContent: true);
     await _prepare();
     setState(() {});
   }
 
   Future<bool> _checkPermission() async {
-    if (widget.platform == TargetPlatform.android) {
+    if (Platform.isIOS) return true;
+
+    DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+    AndroidDeviceInfo androidInfo = await deviceInfo.androidInfo;
+    if (widget.platform == TargetPlatform.android &&
+        androidInfo.version.sdkInt <= 28) {
       final status = await Permission.storage.status;
       if (status != PermissionStatus.granted) {
         final result = await Permission.storage.request();
@@ -335,35 +354,35 @@ class _MyHomePageState extends State<MyHomePage> {
     _tasks = [];
     _items = [];
 
-    _tasks.addAll(_documents.map((document) =>
+    _tasks!.addAll(_documents.map((document) =>
         _TaskInfo(name: document['name'], link: document['link'])));
 
     _items.add(_ItemHolder(name: 'Documents'));
-    for (int i = count; i < _tasks.length; i++) {
-      _items.add(_ItemHolder(name: _tasks[i].name, task: _tasks[i]));
+    for (int i = count; i < _tasks!.length; i++) {
+      _items.add(_ItemHolder(name: _tasks![i].name, task: _tasks![i]));
       count++;
     }
 
-    _tasks.addAll(_images
+    _tasks!.addAll(_images
         .map((image) => _TaskInfo(name: image['name'], link: image['link'])));
 
     _items.add(_ItemHolder(name: 'Images'));
-    for (int i = count; i < _tasks.length; i++) {
-      _items.add(_ItemHolder(name: _tasks[i].name, task: _tasks[i]));
+    for (int i = count; i < _tasks!.length; i++) {
+      _items.add(_ItemHolder(name: _tasks![i].name, task: _tasks![i]));
       count++;
     }
 
-    _tasks.addAll(_videos
+    _tasks!.addAll(_videos
         .map((video) => _TaskInfo(name: video['name'], link: video['link'])));
 
     _items.add(_ItemHolder(name: 'Videos'));
-    for (int i = count; i < _tasks.length; i++) {
-      _items.add(_ItemHolder(name: _tasks[i].name, task: _tasks[i]));
+    for (int i = count; i < _tasks!.length; i++) {
+      _items.add(_ItemHolder(name: _tasks![i].name, task: _tasks![i]));
       count++;
     }
 
-    tasks?.forEach((task) {
-      for (_TaskInfo info in _tasks) {
+    tasks!.forEach((task) {
+      for (_TaskInfo info in _tasks!) {
         if (info.link == task.url) {
           info.taskId = task.taskId;
           info.status = task.status;
@@ -374,12 +393,8 @@ class _MyHomePageState extends State<MyHomePage> {
 
     _permissionReady = await _checkPermission();
 
-    _localPath = (await _findLocalPath()) + Platform.pathSeparator + 'Download';
-
-    final savedDir = Directory(_localPath);
-    bool hasExisted = await savedDir.exists();
-    if (!hasExisted) {
-      savedDir.create();
+    if (_permissionReady) {
+      await _prepareSaveDir();
     }
 
     setState(() {
@@ -387,29 +402,47 @@ class _MyHomePageState extends State<MyHomePage> {
     });
   }
 
-  Future<String> _findLocalPath() async {
-    final directory = widget.platform == TargetPlatform.android
-        ? await getExternalStorageDirectory()
-        : await getApplicationDocumentsDirectory();
-    return directory.path;
+  Future<void> _prepareSaveDir() async {
+    _localPath = (await _findLocalPath())!;
+    final savedDir = Directory(_localPath);
+    bool hasExisted = await savedDir.exists();
+    if (!hasExisted) {
+      savedDir.create();
+    }
+  }
+
+  Future<String?> _findLocalPath() async {
+    var externalStorageDirPath;
+    if (Platform.isAndroid) {
+      try {
+        externalStorageDirPath = await AndroidPathProvider.downloadsPath;
+      } catch (e) {
+        final directory = await getExternalStorageDirectory();
+        externalStorageDirPath = directory?.path;
+      }
+    } else if (Platform.isIOS) {
+      externalStorageDirPath =
+          (await getApplicationDocumentsDirectory()).absolute.path;
+    }
+    return externalStorageDirPath;
   }
 }
 
 class DownloadItem extends StatelessWidget {
-  final _ItemHolder data;
-  final Function(_TaskInfo) onItemClick;
-  final Function(_TaskInfo) onAtionClick;
+  final _ItemHolder? data;
+  final Function(_TaskInfo?)? onItemClick;
+  final Function(_TaskInfo)? onActionClick;
 
-  DownloadItem({this.data, this.onItemClick, this.onAtionClick});
+  DownloadItem({this.data, this.onItemClick, this.onActionClick});
 
   @override
   Widget build(BuildContext context) {
     return Container(
       padding: const EdgeInsets.only(left: 16.0, right: 8.0),
       child: InkWell(
-        onTap: data.task.status == DownloadTaskStatus.complete
+        onTap: data!.task!.status == DownloadTaskStatus.complete
             ? () {
-                onItemClick(data.task);
+                onItemClick!(data!.task);
               }
             : null,
         child: Stack(
@@ -422,7 +455,7 @@ class DownloadItem extends StatelessWidget {
                 children: <Widget>[
                   Expanded(
                     child: Text(
-                      data.name,
+                      data!.name!,
                       maxLines: 1,
                       softWrap: true,
                       overflow: TextOverflow.ellipsis,
@@ -430,33 +463,33 @@ class DownloadItem extends StatelessWidget {
                   ),
                   Padding(
                     padding: const EdgeInsets.only(left: 8.0),
-                    child: _buildActionForTask(data.task),
+                    child: _buildActionForTask(data!.task!),
                   ),
                 ],
               ),
             ),
-            data.task.status == DownloadTaskStatus.running ||
-                    data.task.status == DownloadTaskStatus.paused
+            data!.task!.status == DownloadTaskStatus.running ||
+                    data!.task!.status == DownloadTaskStatus.paused
                 ? Positioned(
                     left: 0.0,
                     right: 0.0,
                     bottom: 0.0,
                     child: LinearProgressIndicator(
-                      value: data.task.progress / 100,
+                      value: data!.task!.progress! / 100,
                     ),
                   )
                 : Container()
-          ].where((child) => child != null).toList(),
+          ].toList(),
         ),
       ),
     );
   }
 
-  Widget _buildActionForTask(_TaskInfo task) {
+  Widget? _buildActionForTask(_TaskInfo task) {
     if (task.status == DownloadTaskStatus.undefined) {
       return RawMaterialButton(
         onPressed: () {
-          onAtionClick(task);
+          onActionClick!(task);
         },
         child: Icon(Icons.file_download),
         shape: CircleBorder(),
@@ -465,7 +498,7 @@ class DownloadItem extends StatelessWidget {
     } else if (task.status == DownloadTaskStatus.running) {
       return RawMaterialButton(
         onPressed: () {
-          onAtionClick(task);
+          onActionClick!(task);
         },
         child: Icon(
           Icons.pause,
@@ -477,7 +510,7 @@ class DownloadItem extends StatelessWidget {
     } else if (task.status == DownloadTaskStatus.paused) {
       return RawMaterialButton(
         onPressed: () {
-          onAtionClick(task);
+          onActionClick!(task);
         },
         child: Icon(
           Icons.play_arrow,
@@ -497,7 +530,7 @@ class DownloadItem extends StatelessWidget {
           ),
           RawMaterialButton(
             onPressed: () {
-              onAtionClick(task);
+              onActionClick!(task);
             },
             child: Icon(
               Icons.delete_forever,
@@ -518,7 +551,7 @@ class DownloadItem extends StatelessWidget {
           Text('Failed', style: TextStyle(color: Colors.red)),
           RawMaterialButton(
             onPressed: () {
-              onAtionClick(task);
+              onActionClick!(task);
             },
             child: Icon(
               Icons.refresh,
@@ -538,19 +571,19 @@ class DownloadItem extends StatelessWidget {
 }
 
 class _TaskInfo {
-  final String name;
-  final String link;
+  final String? name;
+  final String? link;
 
-  String taskId;
-  int progress = 0;
-  DownloadTaskStatus status = DownloadTaskStatus.undefined;
+  String? taskId;
+  int? progress = 0;
+  DownloadTaskStatus? status = DownloadTaskStatus.undefined;
 
   _TaskInfo({this.name, this.link});
 }
 
 class _ItemHolder {
-  final String name;
-  final _TaskInfo task;
+  final String? name;
+  final _TaskInfo? task;
 
   _ItemHolder({this.name, this.task});
 }